refactor(drive): version the shape lowerings behind their own method-version slots
The v1 non-primary-key lowering routed by shape to two unversioned
worker methods. Give each shape lowering the standard versioned-method
structure: get_non_primary_key_single_in_path_query and
get_non_primary_key_multiple_in_path_query dispatch through new
DriveDocumentQueryMethodVersions slots (0 in every table) to frozen
v0 modules holding the constructions and their recursion helpers, so a
future change to one shape bumps that slot alone instead of copying the
whole v1 lowering. Unknown slot values error with
UnknownVersionMismatch. No behavior change.
