+++ packages/rs-platform-version/src/version/drive_versions/drive_document_method_versions/v2.rs
@@ -21,6 +21,8 @@ pub const DRIVE_DOCUMENT_METHOD_VERSIONS_V2: DriveDocumentMethodVersions =
             detect_sum_mode: 0,
             detect_ranked_mode: 0,
             non_primary_key_path_query: 0,
+            non_primary_key_single_in_path_query: 0,
+            non_primary_key_multiple_in_path_query: 0,
             where_clause_grouping: 0,
         },
         delete: DriveDocumentDeleteMethodVersions {
