+++ packages/rs-platform-version/src/version/drive_versions/drive_document_method_versions/v4.rs
@@ -78,6 +78,8 @@ pub const DRIVE_DOCUMENT_METHOD_VERSIONS_V4: DriveDocumentMethodVersions =
             detect_sum_mode: 0,
             detect_ranked_mode: 0,
             non_primary_key_path_query: 1,
+            non_primary_key_single_in_path_query: 0,
+            non_primary_key_multiple_in_path_query: 0,
             where_clause_grouping: 1,
         },
         delete: DriveDocumentDeleteMethodVersions {
